refactor: extract classes to them file and refactor query string reader
