--- src/Javid Hassani/PipelineDesignPattern/PipelineDesignPattern/Controllers/ControllerBase.cs
@@ -1,3 +1,0 @@
-﻿namespace PipelineDesignPattern.Controllers;
-
-public class ControllerBase(Context context);
